Crate hdk

File holding the public Zome API All API Reference documentation should be done here.

Re-exports

pub extern crate serde;

pub extern crate holochain_wasm_utils;

Modules

global_fns

globals
File for holding all internal/private globals used by the zome api library

init_globals
File for holding the internal/private zome api function init_globals

macros

Macros

load_json

validations

zome_functions
A macro for easily writing zome functions

Structs

APP_AGENT_ID_STR
The identity string used when the chain was first initialized. If you used JSON to embed multiple properties (such as FirstName, LastName, Email, etc), they can be retrieved here as App.Agent.FirstName, etc. (FIXME)

APP_AGENT_INITIAL_HASH
The hash of the first identity entry on your chain (The second entry on your chain). This is your peer's identity on the DHT.

APP_AGENT_KEY_HASH
The hash of your public key. This is your node address on the DHT. It can be used for node-to-node messaging with send and receive functions.

APP_AGENT_LATEST_HASH
The hash of the most recent identity entry that has been committed to your chain. Starts with the same value as APP_AGENT_INITIAL_HASH. After a call to update_agent it will have the value of the hash of the newly committed identity entry.

APP_DNA_HASH
The hash of this Holochain's DNA. Nodes must run the same DNA to be on the same DHT.

APP_NAME
The name of this Holochain taken from its DNA.

EntryStatus

GetEntryMask

ValidationData

Enums

BundleOnClose
Allowed input for close_bundle()

EntryAction

EntryLifecycle

LinkAction

RibosomeError

Functions

call
FIXME DOC

close_bundle
FIXME DOC

commit_entry
FIXME DOC

debug
FIXME DOC

get_entry
implements access to low-level WASM hc_get_entry

get_links
FIXME DOC

link_entries
FIXME DOC

make_hash
FIXME DOC

property
FIXME DOC Returns an application property, which are defined by the app developer. It returns values from the DNA file that you set as properties of your application (e.g. Name, Language, Description, Author, etc.).

query
FIXME DOC

remove_entry
FIXME DOC Commit a Deletion System Entry

send
FIXME DOC

sign
FIXME DOC

start_bundle
FIXME DOC

update_agent
FIXME DOC

update_entry
FIXME DOC

verify_signature
FIXME DOC