Struct hdk::entry_definition::ValidatingLinkDefinition

pub struct ValidatingLinkDefinition {
    pub link_type: LinkDirection,
    pub other_entry_type: String,
    pub tag: String,
    pub package_creator: PackageCreator,
    pub validator: LinkValidator,
}

Similar to ValidatingEntryType, this provides the dynamic aspects of link definitions, the validation callbacks, and thus completes the structs in the DNA crate. The entry! macro expects an array of links that are represented by instances of this struct.

DNA developers don't need to use this type directly but instead should use the link!, to! or from! macro.

Fields
link_type: LinkDirection

Is this link defined as pointing from this entry type to some other type, or from the other type to this?

other_entry_type: String

The other entry type the link connects this entry type to

tag: String

Tag (i.e. name) of this type of links

package_creator: PackageCreator

Callback that returns a validation package definition that Holochain reads in order to create the right validation package to pass in to the validator callback on validation.

validator: LinkValidator

This is the validation callback that is used to determine if a link is valid.