Module hdk::meta

This file contains the "secret" functions that get added to Zomes, by the HDK. These functions match expectations that Holochain has... every Zome technically needs these functions, but not every developer should have to write them. A notable function defined here is __hdk_get_json_definition which allows Holochain to retrieve JSON defining the Zome.

Structs

ZomeDefinition

Functions

__hdk_get_json_definition

__hdk_get_validation_package_for_entry_type

__hdk_get_validation_package_for_link

__hdk_git_hash

__hdk_validate_app_entry

__hdk_validate_link