Module hdk::utils

Functions

commit_and_link
Commits the given entry and links it from the base with the given link_type.

get_as_type
Helper function for loading an entry and converting to a given type

get_links_and_load_type
Helper function that perfoms a try_from for every entry of a get_links_and_load for a given type. Any entries that either fail to load or cannot be converted to the type will be dropped.

link_entries_bidir
Creates two links: From A to B, and from B to A, with given link_types.